refactor: move dialogs/symbology to features/layers/symbology
