+++ packages/base/src/commands/index.ts
@@ -22,8 +22,8 @@ import { addLayerCreationCommands } from './operationCommands';
 import { CommandIDs, icons } from '../constants';
 import { ProcessingFormDialog } from '../dialogs/ProcessingFormDialog';
 import { LayerCreationFormDialog } from '../dialogs/layerCreationFormDialog';
-import { SymbologyWidget } from '../dialogs/symbology/symbologyDialog';
 import { LayerBrowserWidget } from '../features/layer-browser';
+import { SymbologyWidget } from '../features/layers/symbology/symbologyDialog';
 import { getSingleSelectedLayer } from '../features/processing/index';
 import { addProcessingCommands } from '../features/processing/processingCommands';
 import { targetWithCenterIcon } from '../icons';
