+++ packages/base/src/features/layers/symbology/classificationModes.ts
