+++ packages/base/src/features/layers/symbology/components/color_ramp/ModeSelectRow.tsx
