+++ packages/base/src/features/layers/symbology/hooks/useEffectiveSymbologyParams.ts
