+++ packages/base/src/features/layers/symbology/hooks/useGetBandInfo.ts
