+++ packages/base/src/features/layers/symbology/hooks/useGetProperties.ts
