+++ packages/base/src/features/layers/symbology/symbologyDialog.tsx
