+++ packages/base/src/features/layers/symbology/symbologyUtils.ts
