+++ packages/base/src/features/layers/symbology/vector_layer/components/ValueSelect.tsx
