+++ packages/base/src/formbuilder/objectform/components/SegmentFormSymbology.tsx
@@ -3,7 +3,7 @@ import { ArrayFieldTemplateProps } from '@rjsf/core';
 import { ChevronRightIcon } from 'lucide-react';
 import React from 'react';
 
-import { SymbologyWidget } from '@/src/dialogs/symbology/symbologyDialog';
+import { SymbologyWidget } from '@/src/features/layers/symbology/symbologyDialog';
 import { Button } from '@/src/shared/components/Button';
 import {
   Collapsible,
